move ginkgo dist implementation
